Extract Security Test Setup
- SecurityTest annotation for Test Setup
- SecurityTestController for testing with Spring Security
