+++ sagan-site/src/test/java/sagan/SecurityConfigTests.java
@@ -1,21 +1,11 @@
 package sagan;
 
-import sagan.security.GithubMemberOAuth2UserService;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.security.oauth2.client.servlet.OAuth2ClientAutoConfiguration;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Import;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -23,15 +13,14 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.head;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 /**
  * @author Rob Winch
  */
-
+@SecurityTest
 @RunWith(SpringRunner.class)
-@WebMvcTest(controllers = SecurityConfigTests.TestController.class)
-@TestPropertySource(properties = "spring.profiles.active=standalone")
 public class SecurityConfigTests {
 	@Autowired
 	MockMvc mockMvc;
@@ -149,13 +138,4 @@ public class SecurityConfigTests {
 				.andExpect(status().isOk());
 	}
 
-	@RestController
-	@TestConfiguration
-	@Import({OAuth2ClientAutoConfiguration.class, GithubMemberOAuth2UserService.class})
-	static class TestController {
-		@RequestMapping("/**")
-		String ok() {
-			return "ok";
-		}
-	}
 }
